Add helper for badge type
